requirements.txt
# Core
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
python-multipart
# Visione (compatibile con lavis)
Pillow==10.3.0

# API
fastapi==0.111.0
uvicorn==0.30.0
requests==2.32.2

# LAVIS (porta con sé opencv-python-headless 4.5.5.64)
salesforce-lavis==1.0.2

